+++ app.js
@@ -2,12 +2,8 @@ const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
 
-//
-const fs = require("fs").promises;
-const path = require("path");
-//
-
 const contactsRouter = require("./routes/api/contacts");
+const { isContactIdExist } = require("./middleware/isIdExist");
 
 const app = express();
 
@@ -19,34 +15,7 @@ app.use(express.json());
 
 // middleWare
 // checks if contact exist
-const contactsPath = path.resolve("./models/contacts.json");
-const getParsedContacts = async () => {
-  try {
-    const data = await fs.readFile(contactsPath);
-    return JSON.parse(data);
-  } catch (error) {
-    console.log(error);
-  }
-};
-
-app.use("/api/contacts/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const contactsData = await getParsedContacts();
-
-    // is array contains id
-    const isIdExist = contactsData.find(
-      (item) => item.id === contactId.toString()
-    );
-    if (!isIdExist) {
-      return res.status(404).json({ message: "Not found" });
-    }
-    next();
-  } catch (error) {
-    console.log(error);
-  }
-});
-// End of middleWare
+app.use("/api/contacts/:contactId", isContactIdExist);
 
 app.use("/api/contacts", contactsRouter);
