+++ helpers/tryCatchHelper.js
@@ -1,0 +1,9 @@
+const asyncWrapper = (controller) => {
+  return (req, res, next) => {
+    controller(req, res).catch(next);
+  };
+};
+
+module.exports = {
+  asyncWrapper,
+};
