+++ routes/api/contacts.js
@@ -1,93 +1,30 @@
 const express = require("express");
-const Joi = require("joi");
-
+// controllers //
 const {
-  listContacts,
-  addContact,
-  getContactById,
-  removeContact,
-  updateContact,
-} = require("../../models/contacts");
+  getContacts,
+  getOneContactById,
+  postContact,
+  deleteContact,
+  putContact,
+} = require("../../controllers/contactsControllers");
+const { asyncWrapper } = require("../../helpers/tryCatchHelper");
+
+// middleware //
+const {
+  postContactValidation,
+  putContactValidation,
+} = require("../../middleware/validationMdlw");
 
 const router = express.Router();
 
-router.get("/", async (req, res, next) => {
-  const contactsList = await listContacts();
-  // res.status(200).json(JSON.stringify(contactsList));
-  res.status(200).json(contactsList);
-});
+router.get("/", asyncWrapper(getContacts));
 
-router.get("/:contactId", async (req, res, next) => {
-  const { contactId } = req.params;
-  const filteredContact = await getContactById(contactId);
+router.get("/:id", asyncWrapper(getOneContactById));
 
-  res.status(200).json(filteredContact);
-});
+router.post("/", postContactValidation, asyncWrapper(postContact));
 
-router.post("/", async (req, res, next) => {
-  // Joi
-  const schema = Joi.object({
-    name: Joi.string().min(3).max(30).required(),
-    email: Joi.string()
-      .email({
-        minDomainSegments: 2,
-        // tlds: { allow: ["com", "net"] },
-      })
-      .required(),
-    phone: Joi.string()
-      .min(3)
-      .max(15)
-      .pattern(/^[0-9-()+ ]+$/)
-      .required(),
-  });
-  const valodationResult = schema.validate(req.body);
+router.delete("/:id", asyncWrapper(deleteContact));
 
-  if (valodationResult.error) {
-    return res.status(400).json({ message: "missing required name field" });
-  }
-  // Joi end
-
-  const { id, name, email, phone } = await addContact(req.body);
-
-  res.status(201).json({
-    id,
-    name,
-    email,
-    phone,
-  });
-});
-
-router.delete("/:contactId", async (req, res, next) => {
-  const { contactId } = req.params;
-  await removeContact(contactId);
-
-  res.status(200).json({ message: "contact deleted" });
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  const { contactId } = req.params;
-
-  // Joi
-  const schema = Joi.object({
-    name: Joi.string().min(3).max(30),
-    email: Joi.string().email({
-      minDomainSegments: 2,
-    }),
-    phone: Joi.string()
-      .min(3)
-      .max(15)
-      .pattern(/^[0-9-()+ ]+$/),
-  }).min(1);
-  const valodationResult = schema.validate(req.body);
-
-  if (valodationResult.error) {
-    return res.status(400).json({ message: "missing fields" });
-  }
-  // Joi end
-
-  const updatedContact = await updateContact(contactId, req.body);
-
-  res.status(200).json(updatedContact);
-});
+router.put("/:id", putContactValidation, asyncWrapper(putContact));
 
 module.exports = router;
